Добавит проверку на наличие изображения с указанным ИД при удалении

+++ src/api/deleteImage.js
@@ -1,8 +1,13 @@
 import imageService from '../service/imageService.js';
 
 const deleteImage = (req, res) => {
-    imageService.remove(req.params.id);
-    res.end();
+    if (imageService.exists(id)) {
+        imageService.remove(req.params.id);
+        res.end();
+    } else {
+        res.status(400).send('В запросе указан некорректный идентификатор')
+    }
+
 };
 
 export { deleteImage };
